+++ Pocket Casts TV App/UI/Auth/QRCodeView.swift
@@ -52,7 +52,6 @@ struct QRCodeView: View {
     }
 }
 
-
 #Preview {
     QRCodeView(url: "https://pocketcasts.net/pair")
 }
